requirements.txt
Django>=3.0,<3.1
djangorestframework>=3.10.3,<3.11.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0

flake8<=3.6.0,<3.7.0
